refactor gateway methods
prettier
