+++ src/routes/explorer/transaction/[transaction].svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import TransactionBox from '@components/transaction-box/TransactionBox.svelte'
-  import type { Transaction } from '@types'
+  import type { Transaction } from './[transaction]'
 
   export let tx: Transaction
 </script>
